+++ TPI/ispi_oliveros_gestion/academico/models.py
@@ -1,23 +1,7 @@
 from django.db import models
 from alumnos.models import Alumno
 from django.contrib.auth.models import User
-
-
-class Docente(models.Model):
- 
-    # Añadimos null=True y blank=True
-    user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='perfil_docente', null=True, blank=True)
-
-    dni = models.CharField(max_length=20, unique=True)
-    nombre = models.CharField(max_length=100)
-    apellido = models.CharField(max_length=100)
-    email = models.EmailField(unique=True)
-    telefono = models.CharField(max_length=20, blank=True, null=True)
-
-    def __str__(self):
-        return f"{self.apellido}, {self.nombre}"
-    
-
+from core.models import Docente
 
 class Carrera(models.Model):
     TIPOS_TITULACION = [
